Adeguamento update community
