+++ arena_network/serializers.py
@@ -1,4 +1,6 @@
-from rest_framework import serializers
+from rest_framework import serializers, status
+from rest_framework.response import Response
+
 from .constants import PROHIBITED_WORDS_EN, PROHIBITED_WORDS_IT
 from .models import Community
 from .utils import contains_prohibited_words
@@ -17,5 +19,24 @@ class CommunitySerializer(serializers.ModelSerializer):
 
         if contains_prohibited_words(value, PROHIBITED_WORDS_EN) or contains_prohibited_words(value, PROHIBITED_WORDS_IT):
             raise serializers.ValidationError("Il nome contiene parole proibite.")
+
+        return value
+
+
+class CommunityUpdateSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Community
+        fields = ['bio', 'avatar']
+
+    def validate_bio(self, value):
+
+        if value:
+            value_lower = value.lower()
+            prohibited_words = [word.lower() for word in PROHIBITED_WORDS_IT + PROHIBITED_WORDS_EN]
+
+            for word in prohibited_words:
+                if word in value_lower:
+                    # Solleva un ValidationError
+                    raise serializers.ValidationError("La bio contiene una parola proibita.")
 
         return value
